订单

+++ mktcenter-serviceimpl/src/main/java/com/bizvane/mktcenterserviceimpl/mappers/MktCouponIntegralExchangePOMapper.java
@@ -110,5 +110,5 @@ public interface MktCouponIntegralExchangePOMapper {
     List<MktCouponIntegralExchangeVO> getCouponIntegralExchangeList(MktCouponIntegralExchangeVO vo);
 
     List<MktCouponIntegralExchangeVO> getExchangeLists(CouponRecordVO vo);
-    List<Integer> getExchangeIds(CouponRecordVO vo);
+    List<Long> getExchangeIds(CouponRecordVO vo);
 }
